requirements.txt
fastapi
uvicorn
slack_sdk
python-dotenv
python-multipart